requirements.txt
eth-utils
pillow
web3
